Producto con permisos añadido

+++ src/main/java/net/daw/control/operation/specific/implementation/ProductoControlOperationSpImpl.java
@@ -29,6 +29,7 @@ import net.daw.control.operation.publicinterface.ControlOperationInterface;
 import net.daw.helper.ExceptionBooster;
 import net.daw.helper.FilterBeanHelper;
 import net.daw.helper.ParameterCooker;
+import net.daw.helper.PermissionManager;
 import net.daw.service.specific.implementation.ProductoServiceSpImpl;
 
 public class ProductoControlOperationSpImpl implements ControlOperationInterface {
@@ -36,12 +37,15 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     private ConnectionInterface DataConnectionSource = null;
     private Connection oConnection = null;
     private ProductoServiceSpImpl oProductoService = null;
+    protected boolean perm;
 
     public ProductoControlOperationSpImpl(HttpServletRequest request) throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException, IllegalArgumentException, InvocationTargetException, Exception {
         try {
             DataConnectionSource = new BoneConnectionPoolImpl();
             oConnection = DataConnectionSource.newConnection();
             oProductoService = new ProductoServiceSpImpl(ParameterCooker.prepareObject(request), ParameterCooker.prepareObject(request), oConnection);
+            PermissionManager oPermissionM = new PermissionManager();
+            perm = oPermissionM.getPermission(request, oConnection);
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":ProductoControlOperationSpImpl ERROR: " + ex.getMessage()));
         }
@@ -51,8 +55,12 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String get(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            result = oProductoService.get(ParameterCooker.prepareId(request));
-            closeDB();
+            if (perm) {
+                result = oProductoService.get(ParameterCooker.prepareId(request));
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":get ERROR: " + ex.getMessage()));
         }
@@ -63,8 +71,12 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getaggregateviewone(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            result = oProductoService.getAggregateViewOne(ParameterCooker.prepareId(request));
-            closeDB();
+            if (perm) {
+                result = oProductoService.getAggregateViewOne(ParameterCooker.prepareId(request));
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getaggregateviewone ERROR: " + ex.getMessage()));
         }
@@ -75,8 +87,12 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getprettycolumns(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            result = oProductoService.getPrettyColumns();
-            closeDB();
+            if (perm) {
+                result = oProductoService.getPrettyColumns();
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getprettycolumns ERROR: " + ex.getMessage()));
         }
@@ -87,8 +103,12 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getcolumns(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            result = oProductoService.getColumns();
-            closeDB();
+            if (perm) {
+                result = oProductoService.getColumns();
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getcolumns ERROR: " + ex.getMessage()));
         }
@@ -99,12 +119,16 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getpage(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            Integer intRegsPerPag = ParameterCooker.prepareRpp(request);
-            Integer intPage = ParameterCooker.preparePage(request);
-            ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
-            HashMap<String, String> hmOrder = ParameterCooker.prepareOrder(request);
-            result = oProductoService.getPage(intRegsPerPag, intPage, alFilter, hmOrder);
-            closeDB();
+            if (perm) {
+                Integer intRegsPerPag = ParameterCooker.prepareRpp(request);
+                Integer intPage = ParameterCooker.preparePage(request);
+                ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
+                HashMap<String, String> hmOrder = ParameterCooker.prepareOrder(request);
+                result = oProductoService.getPage(intRegsPerPag, intPage, alFilter, hmOrder);
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getpage ERROR: " + ex.getMessage()));
         }
@@ -115,10 +139,14 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getpages(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            Integer intRegsPerPag = ParameterCooker.prepareRpp(request);
-            ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
-            result = oProductoService.getPages(intRegsPerPag, alFilter);
-            closeDB();
+            if (perm) {
+                Integer intRegsPerPag = ParameterCooker.prepareRpp(request);
+                ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
+                result = oProductoService.getPages(intRegsPerPag, alFilter);
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getpages ERROR: " + ex.getMessage()));
         }
@@ -129,9 +157,13 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getregisters(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
-            result = oProductoService.getCount(alFilter);
-            closeDB();
+            if (perm) {
+                ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
+                result = oProductoService.getCount(alFilter);
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getregisters ERROR: " + ex.getMessage()));
         }
@@ -142,12 +174,16 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String getaggregateviewsome(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            Integer intRegsPerPag = ParameterCooker.prepareRpp(request);
-            Integer intPage = ParameterCooker.preparePage(request);
-            ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
-            HashMap<String, String> hmOrder = ParameterCooker.prepareOrder(request);
-            result = oProductoService.getAggregateViewSome(intRegsPerPag, intPage, alFilter, hmOrder);
-            closeDB();
+            if (perm) {
+                Integer intRegsPerPag = ParameterCooker.prepareRpp(request);
+                Integer intPage = ParameterCooker.preparePage(request);
+                ArrayList<FilterBeanHelper> alFilter = ParameterCooker.prepareFilter(request);
+                HashMap<String, String> hmOrder = ParameterCooker.prepareOrder(request);
+                result = oProductoService.getAggregateViewSome(intRegsPerPag, intPage, alFilter, hmOrder);
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":getaggregateviewsome ERROR: " + ex.getMessage()));
         }
@@ -158,8 +194,12 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String remove(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            result = oProductoService.remove(ParameterCooker.prepareId(request));
-            closeDB();
+            if (perm) {
+                result = oProductoService.remove(ParameterCooker.prepareId(request));
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":remove ERROR: " + ex.getMessage()));
         }
@@ -170,8 +210,12 @@ public class ProductoControlOperationSpImpl implements ControlOperationInterface
     public String set(HttpServletRequest request) throws Exception {
         String result = null;
         try {
-            result = oProductoService.set(ParameterCooker.prepareJson(request));
-            closeDB();
+            if (perm) {
+                result = oProductoService.set(ParameterCooker.prepareJson(request));
+                closeDB();
+            } else {
+                result = "error";
+            }
         } catch (Exception ex) {
             ExceptionBooster.boost(new Exception(this.getClass().getName() + ":set ERROR: " + ex.getMessage()));
         }
